Moved NuGetTargetedPackageManager to Internal Namespace
