+++ Source/Editor/DualityEditor/DualityEditor.csproj
@@ -181,7 +181,7 @@
     <Compile Include="HelpSystem\IReadOnlyHelpInfo.cs" />
     <Compile Include="HelpSystem\IHelpProvider.cs" />
     <Compile Include="PackageManagement\DependencySet.cs" />
-    <Compile Include="PackageManagement\NuGetTargetedPackageManager.cs" />
+    <Compile Include="PackageManagement\Internal\NuGetTargetedPackageManager.cs" />
     <Compile Include="PackageManagement\PackageDependencyWalker.cs" />
     <Compile Include="PackageManagement\PackageCache.cs" />
     <Compile Include="PackageManagement\PackageManagerLogger.cs" />
